Dockerized the whole thing to run under traefik
Dockerfile works to create an image, it's on docker hub at https://hub.docker.com/repository/docker/coverj715/kahoot-tools. This works only at kahoot.jakecover.me because the configuration is hardcoded into the build, so I still need to add a way to load the cors ip at runtime for this to be actually usable by anyone else.